functions/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
mangum